+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v1.1
 - Add Japanese, Traditional Chinese locale files.
 - Adjust layouts of popup.
+- Add in page selection translate, DeepL auth key (free) request.
+- Add an option page to config the key and the target language.
 
 ## v1.0 (2022-07-08)
 - Release.